requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
celery==5.3.4
redis==5.0.1
aiogram==3.2.0
python-dotenv==1.0.0
Pillow==10.1.0
psycopg2-binary==2.9.9
